requirements.txt
flask==1.1.2
sqlalchemy==1.3.15
flask-sqlalchemy==2.4.1
psycopg2-binary==2.8.4
sqlalchemy-utils==0.36.3
flask-debugtoolbar==0.11.0
flask-wtf==0.14.3
Flask-Bcrypt==0.7.1  
Flask-DebugToolbar==0.11.0 
Flask-Login==0.5.0    
gunicorn==20.0.4
python-dotenv==0.14.0
Flask-migrate==2.5.3